Split protected-item models into one-type-per-file
