+++ tools/Azure.Mcp.Tools.AzureBackup/src/Models/ProtectedItemDiskExclusionProperties.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.AzureBackup.Models;
+
+public sealed record ProtectedItemDiskExclusionProperties(
+    IReadOnlyList<int>? DiskLunList,
+    bool? IsInclusionList);
