+++ tools/Azure.Mcp.Tools.AzureBackup/src/Models/ProtectedItemHealthDetails.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.AzureBackup.Models;
+
+public sealed record ProtectedItemHealthDetails(
+    int? Code,
+    string? Title,
+    string? Message,
+    IReadOnlyList<string>? Recommendations);
