+++ tools/Azure.Mcp.Tools.AzureBackup/src/Models/ProtectedItemKpiHealthDetails.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.AzureBackup.Models;
+
+public sealed record ProtectedItemKpiHealthDetails(
+    string? ResourceHealthStatus,
+    IReadOnlyList<ProtectedItemHealthDetails>? ResourceHealthDetails);
